- TextFormFields validators, labelText and conrollers are added in AddItemScreen.dart - Creating an Item with the input given in AddItemScreen.dart and inserting this Item into the database functionality is added.

+++ lib/data/Item.dart
@@ -6,8 +6,8 @@ class Item {
     final int id;
     final String title;
     final int durationInDay;
-    final Float cost;
-    final Float hourlyCost;
+    final double cost;
+    final double hourlyCost;
     final int projectId;
 
   const Item({
